Fix replicasready autoscaler issue (#20)
* Don't check MachinePool.Spec.Replicas for ReplicasReady

Cluster Autoscaler is changing the number of nodes dynamically,
depending on the cluster workload demands. This means that desired
number of replicas that is set in MachinePool.Spec.Replicas does
not always have a correct value.

For now we can ignore MachinePool.Spec.Replicas when setting
ReplicasReady. We can have a better solution in the future once we
have adopted Cluster API even more and when we can start using
CAPI provider for Cluster Autoscaler.

* Update CHANGELOG